@@ -1,6 +1,6 @@
 {
   "name": "elemental",
-  "version": "0.5.11",
+  "version": "0.6.1",
   "description": "React UI Framework",
   "main": "lib/Elemental.js",
   "author": "Jed Watson",
@@ -10,32 +10,31 @@
     "url": "https://github.com/elementalui/elemental.git"
   },
   "dependencies": {
-    "blacklist": "^1.1.2",
-    "classnames": "^2.2.0"
+    "blacklist": "^1.1.4",
+    "classnames": "^2.2.5"
   },
   "devDependencies": {
     "babel": "^5.8.23",
-    "babel-eslint": "^4.1.3",
-    "eslint": "^1.10.3",
-    "eslint-plugin-react": "^3.13.1",
-    "glob": "^5.0.14",
-    "gulp": "^3.9.0",
-    "history": "^1.17.0",
-    "isparta": "^3.0.4",
-    "react": "^0.14.0",
-    "react-addons-css-transition-group": "^0.14.0",
+    "eslint": "^2.13.1",
+    "eslint-config-keystone": "^2.4.0",
+    "eslint-plugin-react": "^5.2.2",
+    "glob": "^7.0.5",
+    "gulp": "^3.9.1",
+    "isparta": "^4.0.0",
+    "react": "^15.0.0",
+    "react-addons-css-transition-group": "^15.0.0",
     "react-component-gulp-tasks": "^0.7.7",
-    "react-dom": "^0.14.0",
-    "react-router": "1.0.3",
-    "rimraf": "^2.4.3",
-    "tap-xunit": "^1.3.1",
-    "tape": "^4.2.0",
-    "teaspoon": "^5.0.1"
+    "react-dom": "^15.0.0",
+    "react-router": "^2.6.0",
+    "rimraf": "^2.5.4",
+    "tap-xunit": "^1.4.0",
+    "tape": "^4.6.0",
+    "teaspoon": "^6.4.1"
   },
   "peerDependencies": {
-    "react": "^0.14.0",
-    "react-dom": "^0.14.0",
-    "react-addons-css-transition-group": "^0.14.0"
+    "react": "^0.14.0 || ^15.0.0",
+    "react-dom": "^0.14.0 || ^15.0.0",
+    "react-addons-css-transition-group": "^0.14.0 || ^15.0.0"
   },
   "browserify-shim": {
     "react": "global:React"
@@ -44,7 +43,7 @@
     "build": "gulp build",
     "build:lib": "gulp build:lib",
     "bump": "gulp bump",
-    "lint": "eslint ./",
+    "lint": "eslint .",
     "prepublish": "npm run build:lib",
     "publish:npm": "gulp publish:npm",
     "publish:site": "gulp publish:examples",
