@@ -45,6 +45,10 @@ const BUTTON_HOLLOW_VARIANTS = [
 const BUTTON_LINK_VARIANTS = [
 	{ label: 'Link',             value: 'link' },
 	{ label: 'Link Text',        value: 'link-text' },
+	{ label: 'Link Primary',     value: 'link-primary' },
+	{ label: 'Link Success',     value: 'link-success' },
+	{ label: 'Link Warning',     value: 'link-warning' },
+	{ label: 'Link Danger',      value: 'link-danger' },
 	{ label: 'Link Cancel',      value: 'link-cancel' },
 	{ label: 'Link Delete',      value: 'link-delete' }
 ];
@@ -77,6 +81,12 @@ var Buttons = React.createClass({
 			);
 		});
 	},
+	renderButtonVariantsSource(variantType) {
+		var parts = variantType.map(type => {
+			return `<Button type="${type.value}">${type.label}</Button>`;
+		});
+		return parts.join('\n');
+	},
 
 	render () {
 		return (
@@ -104,12 +114,7 @@ var Buttons = React.createClass({
 						{this.renderButtonVariants(BUTTON_VARIANTS)}
 					</div>
 					<ExampleSource>
-						{`
-							<Button type="primary">Primary</Button>
-							<Button type="success">Success</Button>
-							<Button type="warning">Warning</Button>
-							<Button type="danger">Danger</Button>
-						`}
+						{this.renderButtonVariantsSource(BUTTON_VARIANTS)}
 					</ExampleSource>
 				</div>
 				<div className="code-example">
@@ -118,12 +123,7 @@ var Buttons = React.createClass({
 						{this.renderButtonVariants(BUTTON_DEFAULT_VARIANTS)}
 					</div>
 					<ExampleSource>
-						{`
-							<Button type="default-primary">Default Primary</Button>
-							<Button type="default-success">Default Success</Button>
-							<Button type="default-warning">Default Warning</Button>
-							<Button type="default-danger">Default Danger</Button>
-						`}
+						{this.renderButtonVariantsSource(BUTTON_DEFAULT_VARIANTS)}
 					</ExampleSource>
 				</div>
 				<div className="code-example">
@@ -132,12 +132,7 @@ var Buttons = React.createClass({
 						{this.renderButtonVariants(BUTTON_HOLLOW_VARIANTS)}
 					</div>
 					<ExampleSource>
-						{`
-							<Button type="hollow-primary">Hollow Primary</Button>
-							<Button type="hollow-success">Hollow Success</Button>
-							<Button type="hollow-warning">Hollow Warning</Button>
-							<Button type="hollow-danger">Hollow Danger</Button>
-						`}
+						{this.renderButtonVariantsSource(BUTTON_HOLLOW_VARIANTS)}
 					</ExampleSource>
 				</div>
 				<div className="code-example">
@@ -146,12 +141,7 @@ var Buttons = React.createClass({
 						{this.renderButtonVariants(BUTTON_LINK_VARIANTS)}
 					</div>
 					<ExampleSource>
-						{`
-							<Button type="link">Link</Button>
-							<Button type="link-text">Link Text</Button>
-							<Button type="link-cancel">Link Cancel</Button>
-							<Button type="link-delete">Link Delete</Button>
-						`}
+						{this.renderButtonVariantsSource(BUTTON_LINK_VARIANTS)}
 					</ExampleSource>
 				</div>
 
